# docs/openapi.yaml
openapi: 3.1.0
info:
  title: eodi.me Engine API
  version: "1.0"
  description: |
    Local HTTP API for the eodi.me City Vibe Engine.

    **Desktop (local):** The engine binds to `127.0.0.1` on a random port.
    The Tauri shell exposes the resolved base URL and a session token via the
    Settings → Local API panel.

    **Server / Docker mode:** Bind to `0.0.0.0:7557` with `--bind=0.0.0.0 --api-key=<hex>`.

    ## Authentication
    All endpoints require a `Bearer` token in the `Authorization` header.
    Two token types are accepted:
    - **Session token** — ephemeral, rotates on engine restart (desktop mode)
    - **API key** — persistent hex string, stable across restarts (Business+)

    ## Plan gates
    Some endpoints require an active paid plan (`pro`).
    Requests to gated endpoints with a `free` plan return **402 Payment Required**.
  contact:
    name: eodi.me Support
    url: https://eodi.me
    email: support@eodi.me
  license:
    name: Proprietary
    url: https://eodi.me/terms.html
  termsOfService: https://eodi.me/terms.html
servers:
  - url: http://localhost:{port}
    description: Local desktop engine
    variables:
      port:
        default: "39721"
        description: Resolved dynamically; read from Settings → Local API
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: HEX
  schemas:
    HexDetail:
      type: object
      description: Full analysis result for a single H3 hexagon
      properties:
        h3_index:
          type: string
          example: "871e31b63ffffff"
        score:
          type: number
          format: float
          description: Overall composite vibe score (0–1)
        radar:
          type: object
          additionalProperties:
            type: number
          description: Per-dimension scores (0–1)
        dominant_vibe:
          type: string
          description: Name of the highest-scoring dimension
        city:
          type: string
        country:
          type: string
        lat:
          type: number
        lng:
          type: number
    HexSummary:
      type: object
      properties:
        h3_index:
          type: string
        score:
          type: number
        dominant_vibe:
          type: string
        lat:
          type: number
        lng:
          type: number
...
    Error:
      type: object
      properties:
        error:
          type: string
security:
  - bearerAuth: []
paths:
...
  /hex/search:
    get:
      summary: Alias for /search
      tags: [Search]
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HexSummary"
  /hex/match:
    get:
      summary: Similarity search — find hexagons similar to a query hex
      tags: [Search]
      parameters:
        - name: h3
          in: query
          required: true
          schema:
            type: string
          example: "871e31b63ffffff"
        - name: k
          in: query
          schema:
            type: integer
            default: 5
            maximum: 20
          description: Number of similar results to return
      responses:
        "200":
          description: Similar hexagons ranked by similarity
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: "#/components/schemas/HexSummary"
                    - type: object
                      properties:
                        similarity:
                          type: number
                          description: Similarity score 0–1
  /hex/viewport:
    get:
      summary: List hexagons within a lat/lng bounding box
      tags: [Search]
      parameters:
        - name: min_lat
          in: query
          required: true
          schema:
            type: number
        - name: max_lat
          in: query
          required: true
          schema:
            type: number
        - name: min_lng
          in: query
          required: true
          schema:
            type: number
        - name: max_lng
          in: query
          required: true
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: integer
            default: 200
      responses:
        "200":
          description: Hexagons in viewport
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HexSummary"
...
  /hex/nearest:
    get:
      summary: Find the nearest hexagon to a lat/lng point
      tags: [Search]
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
        - name: lng
          in: query
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Nearest hex
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HexDetail"
        "404":
          description: No hex found nearby
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /cities:
    get:
      summary: List cities, optionally filtered by country
      tags: [Reference]
      parameters:
        - name: country
          in: query
          schema:
            type: string
          description: ISO country code or name filter
      responses:
        "200":
          description: City list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
...
  /user/hexagons/bulk:
    post:
      summary: Bulk set user overlay data for multiple hexes
      tags: [UserOverlay]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: object
              description: Map of h3_index → overlay payload
      responses:
        "200":
          description: Bulk update result
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    type: integer
...